found model tex coords problem

+++ src/Model/G_ModelLoader.h
@@ -23,7 +23,8 @@ private:
 		// normal
 		int normal;
 		// texCoords
-		ZC_Vec2<ushort> texCoords;
+		// ZC_Vec2<ushort> texCoords;
+		ZC_Vec2<float> texCoords;
 	};
 	struct VertNorm
 	{
